Add Bigtable v2 conformance tests
This is currently just for ReadRows, and replaces read-rows-acceptance-test.json in each language repo.
(Note that although both the old and new files are JSON, the format is different - the new file is the JSON representation of tests.proto in the same directory.)

This commit also refactors validator.sh a bit in terms of where and when the googleapis repo is cloned.